Updated Team Delete to reflect db cascade rule;

+++ backend/src/application/handlers/players/DeletePlayerCommandHandler.ts
@@ -4,6 +4,8 @@ import { err, ok } from "neverthrow";
 import IPlayerRepository from "application/interfaces/IPlayerRepository";
 import ApplicationErrorFactory from "application/errors/ApplicationErrorFactory";
 import VALIDATION_ERROR_CODES from "application/errors/VALIDATION_ERROR_CODES";
+import ITeamRepository from "application/interfaces/ITeamRepository";
+import FilterAllTeamsCriteria from "infrastructure/contracts/FilterAllTeamsCriteria";
 
 export type DeletePlayerCommandResult = ICommandResult<IApplicationError[]>;
 
@@ -19,9 +21,11 @@ export class DeletePlayerCommand implements ICommand<DeletePlayerCommandResult> 
 
 export default class CreateTeamCommandHandler implements IRequestHandler<DeletePlayerCommand, DeletePlayerCommandResult> {
     private readonly _playerRepository: IPlayerRepository;
+    private readonly _teamRepository: ITeamRepository;
 
-    constructor(props: { playerRepository: IPlayerRepository }) {
+    constructor(props: { playerRepository: IPlayerRepository; teamRepository: ITeamRepository; }) {
         this._playerRepository = props.playerRepository;
+        this._teamRepository = props.teamRepository;
     }
 
     async handle(command: DeletePlayerCommand): Promise<DeletePlayerCommandResult> {
@@ -34,6 +38,21 @@ export default class CreateTeamCommandHandler implements IRequestHandler<DeleteP
                     code: VALIDATION_ERROR_CODES.ModelAlreadyExists,
                 }),
             );
+        }
+
+        const criteria = new FilterAllTeamsCriteria({
+            name: null,
+            teamMembershipPlayerId: player.id
+        });
+        const teamMemberships = await this._teamRepository.filterAllAsync(criteria);
+        if (teamMemberships.length > 0) {
+            return err(
+                ApplicationErrorFactory.createSingleListError({
+                    message: `Team memberships linked to player of id "${command.id}" still exist. Delete the team memberships before you delete the player.`,
+                    path: ["_"],
+                    code: VALIDATION_ERROR_CODES.IntegrityError,
+                }),
+            ); 
         }
 
         await this._playerRepository.deleteAsync(player);
